# Thread contention with GMP backend