@ckormanyos 2020-04-01
> whether the precision is actually changing

When I first read the headlines of this post, my gut feeling was, ...
only change the precision if it is actually being changed.

... Which is exactly what the PR suggests to do.

Did the changes in your PR restore the 32 core performance you expected?