@jzmaddock 2020-04-01
Ugh, this opens up a whole can of worms that we need to fix.

1) Yes we should only set the precision when it's changed.
2) We should be using atomic operations for this.
3) It's very temping to declare the default precision thread local, but if we do that child threads will no longer see the same default precision as their parent, and that could break a whole lot of code.
4) However, if we *don't* make the default precision thread-local, then different threads operating at different precisions will definitely interfere with one another with unpredictable results.

Thoughts?