@jzmaddock 2020-04-01
I'm perplexed by this as I'm not seeing the contention - rather the opposite (the race condition which clearly needs fixing).  BTW we do *not* ever change the default default precision of the underlying GMP library.  Do you have a clear idea of where the contention is?