@wlandry 2020-04-01
> Did the changes in your PR restore the 32 core performance you expected?

Yes.