@wlandry 2020-04-01
> I'm perplexed by this as I'm not seeing the contention - rather the opposite (the race condition which clearly needs fixing). BTW we do _not_ ever change the default default precision of the underlying GMP library. Do you have a clear idea of where the contention is?

I was mistaken.  The precision is stored in a `static` in `gmp_float_imp`.  I would guess that gcc is adding some mutexes to make sure that that we do not shoot ourselves in the foot when reading and updating it.