@ckormanyos 2020-04-01
>> Thoughts?

> I'm perplexed by this as I'm not seeing the contention

It might be relatively straightforward. But I am not sure.

I have a few preliminary, non-confirmed thoughts.

Each GMP object has a backend struct (a classical C-style struct) that manages its memory pointer and various other characteristics of the GMP number. In Boost.Multiprecision, we wrap this struct. The low-level precision-changing function in GMP, is already atomic and secured by its own mutexes.

So it seems like 1 and 2 in the can of worms reduce to thread-safe handling of `m_old_prec` and `m_new_prec`. Since most guys are running 32-bit or 64-bit architectures, however, modificatioins of these `int` members are pretty much atomic by the architecture of the CPU, due to the fact that they can be read, written, read-write-exchanged in single CPU cycles.

3 and 4 might fall into the region of handling a static default precision which may or may not need to be changed after static initialization (i.e. the static initialization that runs prior to calling the `main()` subroutine),

So the original PR might just be enough.

I will give this considerable more thought.

Kind regards, Chris