@jzmaddock 2020-04-01
>The low-level precision-changing function in GMP, is already atomic and secured by its own mutexes.

We don't call that though.  We hold our own default precision which is quite separate from GMP's.