@jzmaddock 2020-04-01
I think we need something more like: https://github.com/boostorg/multiprecision/pull/212

@wlandry can you give this a whirl?