@wlandry 2020-04-01
> I think we need something more like: #212
> 
> @wlandry can you give this a whirl?

This should work for my case.  It makes it explicit that reads and writes are atomic.  I think it ends up being identical to my changes for the x86_64 platforms I run on, but I could see it making a difference on other platforms.

However, it still uses a global variable, which is kind of a disaster for multithreaded use.  If thread `A` atomically sets the precision to 100, thread `B` can atomically set it to 200.  Then all of the thread `A`s computations will run at precision=200, not 100.  Fixing this requires making the precision non-static.  But then you would not need to make it atomic either.