@jzmaddock 2020-04-01
>However, it still uses a global variable, which is kind of a disaster for multithreaded use. 

I agree, the issue then is that if code sets the default precision at the start of `main()`, and then spawns a thread or 2, those threads *do not* inherit the same default precision as the main thread.  This is IMO a very important use case?