@ckormanyos 2020-04-01
> I think we need something more like: #212 

I'm not sure, and maybe I am missing something simple... But that seems like a lot of effort to make read/write operations to values of type `unsigned` atomic. Maybe I'm missing something, but read/write/modify operations on `unsigned` are almost always atomic unless you are dealing with something like an 8-bit architecture and unsigned is 16 bits or something like that. Then you will need multiple cycles to read/write/modify `unsigned`. But I think most all platforms using multiprecision can do single-cycle read/write/modify of `unsigned`. But I guess just to be sure, they can be atomics.

The issues regarding the static precision returned as a reference do, in fact, seem like they need to be more fully discussed.

>> The low-level precision-changing function in GMP, is already atomic and secured by its own mutexes.

> We don't call that though. We hold our own default precision which is quite separate from GMP's.

Somewhere we must be calling the low-level functions that actually do the memory allocation/deallocation in the depths of GMP. It is these that should already be secured with their own mutexes, those memory accesses. I was not referring to the class-local members of type `unsigned` that can be modified.

Kind regards, Chris