@ckormanyos 2020-04-01
>>> The low-level precision-changing function in GMP, is already atomic and secured by its own mutexes.

>> We don't call that though. We hold our own default precision which is quite separate from GMP's.

> Somewhere we must be calling the low-level functions that actually do the memory allocation/deallocation in the depths of GMP.

My text was vague. I mean subroutines such as `mpf_set_prec_raw()` should already be thread safe in gmp_threads.