@wlandry 2020-04-01
> I agree, the issue then is that if code sets the default precision at the start of `main()`, and then spawns a thread or 2, those threads _do not_ inherit the same default precision as the main thread. This is IMO a very important use case?

I think that the current situation, where thread `A` can interfere with thread `B`, is not acceptable.
Thinking some more, maybe store the default precision in thread local storage?  If it is not set, then it can read the global value once.  As I understand it, thread local storage is a problem on some architectures, but this might be the best way forward.

With all that said, this still does not cover coroutines.  They do not have their own copies of statics, so they can create similar kinds of trouble.  This kind of argues that all numbers should get their precision either explicitly from constructor arguments or implicitly from other high precision numbers.  That would be a big change.

The root of the problem seems to be that boost::multiprecision is setting and resetting the default precision. Does it really have to do that?