@wlandry 2020-04-03
> > It is possible that the performance problem that I am seeing is not from mutexes, but cache invalidation.
> 
> In my experience, the much more common and (for runtime of significant influence) is more often thread traffic. This involves dispatching threads, stopping or blocking threads, or worst case creating threads. Since you are sort of at the whim of gmp_threads, that is sort of a big unknown.

I do not know what you mean by gmp_threads.  GMP is not creating any threads.  The application creates and destroys threads only a few times.  When it does create threads, it only creates a small number of them (it is a run time parameter, generally set to the number of cores). Most of the work is independent and done within the threads.

If you are curious, the application is [scalar_blocks](https://gitlab.com/bootstrapcollaboration/scalar_blocks).

> In order to shed some light on this, I would do things like...
> 
>     * Run with one single thread with native GMP
> 
>     * Run with 2 or 4 threads both with and without your simple fix in this PR.
> 
>     * Run with significantly more digits, such as twice as many or even ten-fold.
> 
> 
> When doing all these things, you will potentially see a pattern. Does it look like thread traffic, or does it look like a memory barrier effect?

Here are two tables for a medium sized-benchmark.  The output is from /usr/bin/time.

Without fix
```
| Threads | Precision (bits) | User (s) | System (s) | Elapsed (s) | CPU % |
|       1 |              665 |    52.33 |       0.13 |       52.54 |    99 |
|       2 |              665 |    81.02 |       0.16 |       42.12 |   192 |
|       4 |              665 |   133.00 |       0.21 |       35.41 |   376 |
|       8 |              665 |   284.18 |       0.18 |       37.06 |   767 |
|      16 |              665 |   497.90 |       0.24 |       33.11 |  1504 |
|      28 |              665 |   775.97 |       0.38 |       30.99 |  2504 |
|       1 |             1230 |   101.78 |       0.43 |      102.36 |    99 |
|       2 |             1230 |   121.39 |       0.40 |       62.12 |   196 |
|       4 |             1230 |   157.51 |       0.36 |       41.23 |   382 |
|       8 |             1230 |   263.51 |       0.36 |       35.52 |   742 |
|      16 |             1230 |   584.73 |       0.41 |       39.15 |  1494 |
|      28 |             1230 |   790.33 |       0.54 |       31.64 |  2499 |
```
With fix
```
| Threads | Precision (bits) | User (s) | System (s) | Elapsed (s) | CPU % |
|       1 |              665 |    53.07 |       0.17 |       53.45 |    99 |
|       2 |              665 |    52.69 |       0.24 |       27.00 |   195 |
|       4 |              665 |    52.69 |       0.18 |       13.62 |   388 |
|       8 |              665 |    52.71 |       0.18 |       07.25 |   729 |
|      16 |              665 |    52.84 |       0.18 |       04.11 |  1288 |
|      28 |              665 |    53.00 |       0.22 |       02.28 |  2327 |
|       1 |             1230 |   102.13 |       0.25 |      102.48 |    99 |
|       2 |             1230 |   122.15 |       0.31 |       62.44 |   196 |
|       4 |             1230 |   138.33 |       0.28 |       35.93 |   385 |
|       8 |             1230 |   177.68 |       0.36 |       24.13 |   737 |
|      16 |             1230 |   377.57 |       0.33 |       25.78 |  1465 |
|      28 |             1230 |   532.65 |       0.49 |       21.69 |  2457 |
```

The most dramatic difference is between the `Elapsed` time with the smaller precision.  That difference is what caused me to look into this.  For large precision, I would guess that the memory allocator is the bottleneck.  The application creates lots of GMP temporaries.  But that is for another day.

> How many digits are you using in the first place?

665 bits, or 200 decimal digits.

> Are you even changing precision? Or are you just getting traffic at the subroutine that _might_ change precision?

I never change precision.

> As a final point, and this is not really related to the original PR, but relevant nonetheless... I almost never recommend the multithreaded granularity on the big-number level (i.e., within the GMP lib). The reason for this is that thread overhead is likely your first and worst offer with that extremely low-level threading granularity. It is moch more efficient in general to identify entire rows of calculations or possible ways for higher-level parallelization. Think like the rows of a matrix, or an entire line of a loop within a calculation. It is with this higher-level granularity that you will minimize threading traffic and overhead and maximize algorithmic output. Again, this has not much to do with the original PR, but check out the possibilities in your desing.

That is what I already do.  This is not an application that creates a bajillion threads and lets the operating system sort it out.